services/scanner/requirements.txt
flask
fastapi
uvicorn
pyyaml